+++ packages/browser-wallet/src/popup/popupX/pages/prompts/SendTransaction/i18n/en.ts
@@ -3,6 +3,14 @@ const t = {
     sign: 'Sign & Submit',
     signTransaction: '<1>{{dApp}}</1> suggests a transaction',
     signRequest: 'Signature request',
+    transferRequest: 'Transfer Request',
+    tokenUpdate: 'Token Update',
+    method: 'Method',
+    advanced: 'Advanced',
+    account: {
+        balance: 'Balance: ',
+        atDisposal: 'At Disposal: ',
+    },
     errors: {
         missingAccount: 'Missing account address',
         missingKey: 'Missing key for the chosen address',
@@ -27,6 +35,18 @@ const t = {
         version: 'Version',
         tokenId: 'Token Id',
         operations: 'Operations',
+        fee: 'Approximate fee',
+    },
+    operations: {
+        transfer: 'Transfer',
+        mint: 'Mint',
+        burn: 'Burn',
+        addAllowList: 'Add to allow list',
+        removeAllowList: 'Remove from allow list',
+        addDenyList: 'Add to deny list',
+        removeDenyList: 'Remove from deny list',
+        pause: 'Pause',
+        unpause: 'Unpause',
     },
 };